got vis working for visual studio

+++ visualization/SchwarmGUI/SchwarmPacket/packet.h
@@ -20,7 +20,7 @@ namespace Schwarm
     {
     private:
         uint8_t* __data;
-        size_t data_size;
+        uint32_t data_size;
 
         void free(void);
 
@@ -30,28 +30,28 @@ namespace Schwarm
         packet_error internal_decode(void);
 
     public:
-        static constexpr size_t SIZE_ID            = sizeof(uint8_t);
-        static constexpr size_t SIZE_PACKET_LENGTH = sizeof(size_t);
+        static constexpr uint32_t SIZE_ID            = sizeof(uint8_t);
+        static constexpr uint32_t SIZE_PACKET_LENGTH = sizeof(uint32_t);
 
         Packet(void);
         Packet(const Packet&);
         Packet(Packet&&);
         virtual ~Packet(void);
 
-        void allocate(size_t);
-        packet_error set(uint8_t*, size_t = 0);
+        void allocate(uint32_t);
+        packet_error set(uint8_t*, uint32_t = 0);
 
         virtual packet_error encode(void) = 0;
         virtual packet_error decode(void) = 0;
 
         virtual uint8_t id(void)        const noexcept = 0;
-        size_t          size(void)      const noexcept;
+        uint32_t        size(void)      const noexcept;
         const uint8_t*  data(void)      const;
         const uint8_t*  rawdata(void)   const;
-        virtual inline size_t min_size(void) const noexcept {return SIZE_ID + SIZE_PACKET_LENGTH;}
+        virtual inline uint32_t min_size(void) const noexcept {return SIZE_ID + SIZE_PACKET_LENGTH;}
 
         static const uint8_t*   id_ptr(const uint8_t*);
-        static const size_t*    size_ptr(const uint8_t*);
+        static const uint32_t*  size_ptr(const uint8_t*);
         static const uint8_t*   data_ptr(const uint8_t*);
 
         static const char* strerror(packet_error) noexcept;
@@ -126,7 +126,7 @@ namespace Schwarm
 
     public:
         static constexpr uint8_t PACKET_ID = 2;
-        static constexpr size_t ERROR_CODE_SIZE = sizeof(packet_error);
+        static constexpr uint32_t ERROR_CODE_SIZE = sizeof(packet_error);
 
         ErrorPacket(void);
         ErrorPacket(const ErrorPacket&);
@@ -137,7 +137,7 @@ namespace Schwarm
         virtual packet_error decode(void);
 
         virtual inline uint8_t id(void) const noexcept      {return PACKET_ID;}
-        virtual inline size_t min_size(void) const noexcept {return SIZE_ID + SIZE_PACKET_LENGTH + ERROR_CODE_SIZE;}
+        virtual inline uint32_t min_size(void) const noexcept {return SIZE_ID + SIZE_PACKET_LENGTH + ERROR_CODE_SIZE;}
 
         void            set_code(packet_error)  noexcept;
         packet_error    get_code(void)          const noexcept;
@@ -158,16 +158,16 @@ namespace Schwarm
         int vehicle_id;
         bool invert;
         char* filepath;
-        size_t fp_allocsize;
+        uint32_t fp_allocsize;
 
-        void alloc_fp(size_t);
+        void alloc_fp(uint32_t);
         void free_fp(void);
 
     public:
         static constexpr uint8_t PACKET_ID = 3;
-        static constexpr size_t SIZE_NUM_GOALS = sizeof(unsigned int);
-        static constexpr size_t SIZE_INVERT = sizeof(bool);
-        static constexpr size_t SIZE_VEHICLE_ID = sizeof(int);
+        static constexpr uint32_t SIZE_NUM_GOALS = sizeof(unsigned int);
+        static constexpr uint32_t SIZE_INVERT = sizeof(bool);
+        static constexpr uint32_t SIZE_VEHICLE_ID = sizeof(int);
 
         PathGeneratePacket(void);
         PathGeneratePacket(const PathGeneratePacket&);
@@ -178,7 +178,7 @@ namespace Schwarm
         virtual packet_error decode(void);
 
         virtual inline uint8_t id(void) const noexcept      {return PACKET_ID;}
-        virtual inline size_t min_size(void) const noexcept {return SIZE_ID + SIZE_PACKET_LENGTH + SIZE_NUM_GOALS + SIZE_VEHICLE_ID + SIZE_INVERT;}
+        virtual inline uint32_t min_size(void) const noexcept {return SIZE_ID + SIZE_PACKET_LENGTH + SIZE_NUM_GOALS + SIZE_VEHICLE_ID + SIZE_INVERT;}
 
         void            set_num_goals(unsigned int) noexcept;
         unsigned int    get_num_goals(void)         const noexcept;
@@ -191,7 +191,7 @@ namespace Schwarm
 
         void        set_filepath(const char*)   noexcept;
         const char* get_filepath(void)          const noexcept;
-        size_t      filepath_size(void)         const noexcept;
+        uint32_t      filepath_size(void)         const noexcept;
 
         PathGeneratePacket& operator=(const PathGeneratePacket&);
         PathGeneratePacket& operator=(PathGeneratePacket&&);
@@ -210,8 +210,8 @@ namespace Schwarm
 
     public:
         static constexpr uint8_t PACKET_ID = 4;
-        static constexpr size_t SIZE_GOAL_IDX = sizeof(size_t);
-        static constexpr size_t SIZE_VEHICLE_ID = sizeof(int);
+        static constexpr uint32_t SIZE_GOAL_IDX = sizeof(uint32_t);
+        static constexpr uint32_t SIZE_VEHICLE_ID = sizeof(int);
 
         GoalReqPacket(void);
         GoalReqPacket(const GoalReqPacket&);
@@ -222,10 +222,10 @@ namespace Schwarm
         virtual packet_error decode(void);
 
         virtual inline uint8_t id(void) const noexcept      {return PACKET_ID;}
-        virtual inline size_t min_size(void) const noexcept {return SIZE_ID + SIZE_PACKET_LENGTH + SIZE_GOAL_IDX + SIZE_VEHICLE_ID;}
+        virtual inline uint32_t min_size(void) const noexcept {return SIZE_ID + SIZE_PACKET_LENGTH + SIZE_GOAL_IDX + SIZE_VEHICLE_ID;}
 
-        void    set_goal_index(size_t)  noexcept;
-        size_t  get_goal_index(void)    const noexcept;
+        void    set_goal_index(uint32_t)  noexcept;
+        uint32_t  get_goal_index(void)    const noexcept;
 
         void set_vehicle_id(int)    noexcept;
         int  get_vehicle_id(void)   const noexcept;
@@ -247,8 +247,8 @@ namespace Schwarm
 
     public:
         static constexpr uint8_t PACKET_ID = 5;
-        static constexpr size_t SIZE_GOAL = 2 * sizeof(float);
-        static constexpr size_t SIZE_VEHICLE_ID = sizeof(int);
+        static constexpr uint32_t SIZE_GOAL = 2 * sizeof(float);
+        static constexpr uint32_t SIZE_VEHICLE_ID = sizeof(int);
 
         GoalPacket(void);
         GoalPacket(const GoalPacket&);
@@ -259,7 +259,7 @@ namespace Schwarm
         virtual packet_error decode(void);
 
         virtual inline uint8_t id(void) const noexcept      {return PACKET_ID;}
-        virtual inline size_t min_size(void) const noexcept {return SIZE_ID + SIZE_PACKET_LENGTH + SIZE_GOAL + SIZE_VEHICLE_ID;}
+        virtual inline uint32_t min_size(void) const noexcept {return SIZE_ID + SIZE_PACKET_LENGTH + SIZE_GOAL + SIZE_VEHICLE_ID;}
 
         void    set_goal(float, float)  noexcept;
         float   get_goal_x(void)        const noexcept;
